# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Event Ticketing System API",
        "description": "A REST API for an Event Ticketing System built with Go and PostgreSQL",
        "version": "1.0.0",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        }
    },
    "host": "localhost:8005",
    "basePath": "/",
    "schemes": ["http", "https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter the token in the format: Bearer {token}"
        }
    },
    "paths": {
...
        "/api/events/{id}/purchase": {
            "post": {
                "summary": "Purchase ticket for an event",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Event ID"
                    },
                    {
                        "in": "body",
                        "name": "purchase",
                        "description": "Ticket purchase data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TicketPurchase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket purchased successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/tickets": {
            "get": {
                "summary": "Get user's tickets",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user's tickets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Ticket"
                            }
                        }
                    }
                }
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "summary": "Get ticket by ID",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Ticket ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details",
                        "schema": {
                            "$ref": "#/definitions/Ticket"
                        }
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        },
        "/api/tickets/{id}/validate": {
            "post": {
                "summary": "Validate ticket",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Ticket ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket validated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        },
        "/api/events/{id}/attendees": {
            "get": {
                "summary": "Get event attendees",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Event ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of event attendees",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AttendanceLog"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        },
        "/api/events/{id}/attendees/export": {
            "get": {
                "summary": "Export event attendees",
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Event ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file with attendees data"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Ticket": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "Ticket ID"
                },
                "event_id": {
                    "type": "integer",
                    "description": "Associated event ID"
                },
                "user_id": {
                    "type": "integer",
                    "description": "User who purchased the ticket"
                },
                "qr_code": {
                    "type": "string",
                    "description": "QR code for ticket validation"
                },
                "is_validated": {
                    "type": "boolean",
                    "description": "Whether the ticket has been validated"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Purchase timestamp"
                }
            }
        },
        "TicketPurchase": {
            "type": "object",
            "required": ["quantity"],
            "properties": {
                "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Number of tickets to purchase"
                }
            }
        },
        "AttendanceLog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "description": "Attendance log ID"
                },
                "ticket_id": {
                    "type": "integer",
                    "description": "Associated ticket ID"
                },
                "user_id": {
                    "type": "integer",
                    "description": "User ID"
                },
                "event_id": {
                    "type": "integer",
                    "description": "Event ID"
                },
                "validated_at": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Validation timestamp"
                }
            }
        }
    }
}